fix: move utils dir to src folder and refactor imports
