+++ plugins/github-codespaces/src/utils/getGitStatusView.tsx
