+++ plugins/github-codespaces/src/utils/isGithubCodespacesAvailable.ts
